Make Achievements a context
- Move Achievement into /achievement, in line with the context convention
- Rename Achievement to Achievements
- Restructure some modules
- Introduce context module, Habits.Achievements
- Behaviour becomes its own module, nice
