+++ lib/habits_web/controllers/api/v1/achievement_controller.ex
@@ -1,6 +1,8 @@
 defmodule HabitsWeb.API.V1.AchievementController do
   use Habits.Web, :controller
 
+  alias Habits.Achievements
+
   @doc """
   Override action/2 to provide current_account to actions
   """
@@ -19,10 +21,10 @@ defmodule HabitsWeb.API.V1.AchievementController do
       |> Repo.get(habit_id)
 
     conn
-    |> render("index.json", achievements: Achievement.all_for(habit))
+    |> render("index.json", achievements: Achievements.all_for(habit))
   end
   def index(conn, _params, current_account) do
     conn
-    |> render("index.json", achievements: Achievement.all_for(current_account))
+    |> render("index.json", achievements: Achievements.all_for(current_account))
   end
 end
